# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AGC Cheswerta CMS API",
    "version": "1.0.0",
    "description": "Church management system API \u2014 members, givings, attendance, fellowship scheduling, council roster, SMS broadcast, and admin tools.\n\n**Getting started:** call `POST /api/auth/login`, copy the `token` from the response, then click **Authorize** above and paste it in as `<token>` (no need to type \"Bearer \" \u2014 Swagger adds that automatically). Every other request will then include it.\n\n**Not included:** a few blueprints registered in the app (announcements, modules, feature flags, generic settings, audit log, account locks, login attempts, file uploads, generic db query, stats/PDF reports, SMS logs, church info) aren't documented here yet \u2014 ask to have them added and share those route files.\n\n**Server selector:** the dropdown at the top switches between Local and Production \u2014 edit the Production URL in the spec to match your real domain."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local dev server (default Flask port)"
    },
    {
      "url": "https://REPLACE-WITH-YOUR-PRODUCTION-DOMAIN",
      "description": "Production \u2014 replace with your real API domain"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the token from POST /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "You do not have permission to perform this action."
          }
        }
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "CreateUserInput": {
        "type": "object",
        "required": [
          "full_name",
          "username",
          "role",
          "password"
        ],
        "properties": {
          "full_name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "admin",
              "pastor",
              "secretary",
              "treasurer",
              "ministry_leader",
              "lay_leader"
            ]
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8
          }
        }
      },
      "UpdateUserInput": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8
          },
          "is_active": {
            "type": "boolean"
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "admin",
              "pastor",
              "secretary",
              "treasurer",
              "ministry_leader",
              "lay_leader"
            ]
          }
        }
      },
...
      "TestSmsInput": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "example": "0712345678"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a user (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Username taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete your own account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rpc/is_account_locked": {
      "post": {
        "tags": [
          "RPC"
        ],
        "summary": "Check whether an identifier is currently lockout-blocked",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_identifier": {
                    "type": "string",
                    "example": "jdoe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lock status"
          }
        }
      }
    },
    "/api/rpc/get_email_by_username": {
      "post": {
        "tags": [
          "RPC"
        ],
        "summary": "Look up a synthetic email by username",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "_username": {
                    "type": "string",
                    "example": "jdoe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email (or null)"
          }
        }
      }
    },
    "/api/rpc/test_sms": {
      "post": {
        "tags": [
          "RPC"
        ],
        "summary": "Send a test SMS using the configured provider",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSmsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sent"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}